+++ lib/flutter_monorepo.dart
@@ -1,5 +1,6 @@
 export 'src/doctor.dart';
 export 'src/generator.dart';
+export 'src/workflow.dart';
 export 'src/project_config.dart';
 export 'src/templates/app/app_template_factory.dart';
 export 'src/templates/app/app_template_strategy.dart' show AppTemplateStrategy, appRoutes;